Backport sql.isEquivalent from pg-sql2@5
